app/3_response_synthesizer_api/requirements/common.txt
faiss-cpu
google-cloud-storage
google-cloud-logging
grpcio
langchain
llama_index==0.9.11
pika
pypdf
docx2txt
sentence-transformers
cryptography>=3.1
transformers
peft
fastapi
uvicorn
gunicorn
psycopg2-binary==2.9.9
pgvector==0.2.5
asyncpg==0.29.0
SQLAlchemy==2.0.29
boto3==1.34.105
